management/requirements.txt
asgiref==3.8.1
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pytest==8.3.2
pytest-django==4.8.0
sqlparse==0.5.1
